Allow specification of CA cert for HTTP status check (#418)
* add ca_certs configuration

* test for ca_certs line

* fix scoping bug

previously, the ca_certs conditional was nested in the 'tls cert
check is disabled' conditional.  while it doesn't make a _ton_ of sense
to provide a ca cert when you aren't checking it, allowing users to do
so makes migrating from an unchecked to a checked workflow easier.  I
don't see a downside to allowing this behavior.

* whitespace to trigger build

* Revert "whitespace to trigger build"

This reverts commit e875f27e3dec57b88b58432b984b4accd84799dd.